pedido-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-pubsub==2.21.1
SQLAlchemy==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
